backend/requirements.txt
Django
djangorestframework
django-filter
django-cors-headers
drf-spectacular
pylint-django
python-dotenv
sentry-sdk[django]
psycopg2
PyJWT[crypto]
svix
requests
clerk-backend-api
azure-storage-blob==12.19.0

# Test dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
freezegun==1.5.1
responses==0.24.1
pytest-xdist==3.5.0